- Rename project - install instructions
