+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.13)
 
-project(custom_urdf VERSION 0.1.0)
+project(brl_urdf VERSION 0.1.0)
 
 message(STATUS "=============  !BRL Custom URDF Parser!  =============")
 
@@ -13,8 +13,6 @@ option(BUILD_TESTS "Building the tests" ON)
 # Setup Compilation
 ################################################################################
 
-# TODO(@MatthewChignoli): How much of this is actually needed?
-
 message(STATUS "======> Setup Compilation")
 
 set(CMAKE_DISABLE_IN_SOURCE_BUILD ON)
@@ -25,7 +23,7 @@ if ("${CMAKE_SOURCE_DIR}" STREQUAL "${CMAKE_BINARY_DIR}")
 endif ()
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
-set(CMAKE_COLOR_MAKEFILE   ON)
+set(CMAKE_COLOR_MAKEFILE ON)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 
@@ -69,38 +67,38 @@ else()
   set(LIBRARY_NAME ${PROJECT_NAME})
   set_target_properties(${PROJECT_NAME} PROPERTIES POSITION_INDEPENDENT_CODE TRUE) 
 
-#   install(TARGETS ${PROJECT_NAME}
-#           EXPORT ${PROJECT_NAME}Targets
-#           LIBRARY DESTINATION lib
-#           ARCHIVE DESTINATION lib
-#           RUNTIME DESTINATION bin
-#           INCLUDES DESTINATION include)
+  install(TARGETS ${PROJECT_NAME}
+          EXPORT ${PROJECT_NAME}Targets
+          LIBRARY DESTINATION lib
+          ARCHIVE DESTINATION lib
+          RUNTIME DESTINATION bin
+          INCLUDES DESTINATION include)
 
-#   install(DIRECTORY include/ DESTINATION include)
+  install(DIRECTORY include/ DESTINATION include)
 
-#   include(CMakePackageConfigHelpers)
-#   write_basic_package_version_file(
-#     ${PROJECT_NAME}ConfigVersion.cmake
-#     VERSION ${PACKAGE_VERSION}
-#     COMPATIBILITY AnyNewerVersion
-#   )
+  include(CMakePackageConfigHelpers)
+  write_basic_package_version_file(
+    ${PROJECT_NAME}ConfigVersion.cmake
+    VERSION ${PACKAGE_VERSION}
+    COMPATIBILITY AnyNewerVersion
+  )
 
-#   configure_package_config_file(
-#     ${PROJECT_SOURCE_DIR}/cmake/${PROJECT_NAME}Config.cmake.in
-#     ${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
-#     INSTALL_DESTINATION lib/cmake/${PROJECT_NAME}
-#   )
+  configure_package_config_file(
+    ${PROJECT_SOURCE_DIR}/cmake/${PROJECT_NAME}Config.cmake.in
+    ${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
+    INSTALL_DESTINATION lib/cmake/${PROJECT_NAME}
+  )
 
-#   install(EXPORT ${PROJECT_NAME}Targets
-#     FILE ${PROJECT_NAME}Targets.cmake
-#     NAMESPACE ${PROJECT_NAME}::
-#     DESTINATION lib/cmake/${PROJECT_NAME})
+  install(EXPORT ${PROJECT_NAME}Targets
+    FILE ${PROJECT_NAME}Targets.cmake
+    NAMESPACE ${PROJECT_NAME}::
+    DESTINATION lib/cmake/${PROJECT_NAME})
 
-#   install(FILES
-#     ${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
-#     ${PROJECT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake
-#     DESTINATION lib/cmake/${PROJECT_NAME}
-#   )
+  install(FILES
+    ${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake
+    ${PROJECT_BINARY_DIR}/${PROJECT_NAME}ConfigVersion.cmake
+    DESTINATION lib/cmake/${PROJECT_NAME}
+  )
 endif()
 
 if(BUILD_TESTS)
