+++ cmake/brl_urdfConfig.cmake.in
@@ -1,0 +1,5 @@
+@PACKAGE_INIT@
+
+include("${CMAKE_CURRENT_LIST_DIR}/brl_urdfTargets.cmake")
+
+check_required_components("brl_urdf")
