+++ include/brl_urdf/exception.h
