+++ tests/testExporter.cpp
@@ -2,7 +2,7 @@
 
 #include <cstdio>
 #include "config.h"
-#include "custom_urdf/urdf_parser.h"
+#include "brl_urdf/urdf_parser.h"
 
 using namespace urdf;
